Break logical chunks in small file modules
