+++ index.js
@@ -1,32 +1,8 @@
 #!/usr/bin/env node
 
-const cli = require('commander');
-const execa = require('execa');
-const config = require('./src/config');
-const findBinary = require('./src/find-binary');
-const pkg = require('./package.json');
+const cliInit = require('./src/cli-init');
+const configLoad = require('./src/config-load');
+const runAll = require('./src/run-all');
 
-cli
-  .version(pkg.version, '-v, --version')
-  .parse(process.argv);
-
-const binary = `${__dirname}/bin/git-run-if-changed.sh`;
-
-function runCommandsIfFileChanged(fileToCheck, commands) {
-  const commandsList = Array.isArray(commands) ? commands : [commands];
-  const commandsListResolved = commandsList.map((cmd) => {
-    const { binaryPath, args: binaryArgs } = findBinary(cmd);
-
-    return [binaryPath].concat(binaryArgs).join(' ');
-  });
-  const args = [fileToCheck].concat(commandsListResolved);
-  execa(binary, args).stdout.pipe(process.stdout);
-}
-
-Object.entries(config).forEach(([file, commands]) => {
-  if (commands.length === 0) {
-    return;
-  }
-
-  runCommandsIfFileChanged(file, commands);
-});
+cliInit();
+runAll(configLoad());
