+++ src/cli-init.js
@@ -1,0 +1,8 @@
+const cli = require('commander');
+const pkg = require('../package.json');
+
+module.exports = function cliInit() {
+  cli
+    .version(pkg.version, '-v, --version')
+    .parse(process.argv);
+};
